LPS-112805 Extract common configuration
for both AutoLogin and AuthVerifier
