+++ modules/apps/portal-security/portal-security-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Security API
+Bundle-SymbolicName: com.liferay.portal.security.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.security.configuration
