+++ modules/apps/portal-security/portal-security-api/src/main/resources/com/liferay/portal/security/configuration/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
