+++ modules/apps/portal-security/portal-security-api/src/main/resources/content/Language.properties
@@ -1,0 +1,1 @@
+basic-auth-header-support-configuration-name=Basic Authentication Protocol Support
