+++ modules/apps/portal-security/portal-security-auth-verifier/build.gradle
@@ -8,7 +8,7 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.http.whiteboard", version: "1.0.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
-	compileOnly project(":apps:portal-security:portal-security-auto-login")
+	compileOnly project(":apps:portal-security:portal-security-api")
 	compileOnly project(":apps:portal:portal-upgrade-api")
 	compileOnly project(":apps:static:osgi:osgi-util")
 	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
