+++ modules/apps/portal-security/portal-security-auto-login/src/main/resources/content/Language.properties
@@ -4,4 +4,3 @@ basic-auth-header-auto-login-configuration-name=Auto Login Basic Authentication 
 request-header-auto-login-configuration-name=Auto Login Request Header
 request-parameter-auto-login-configuration-name=Auto Login Request Parameter
 rest.extender.configuration.name=REST Extenders
-basic-auth-header-support-configuration-name=Basic Authentication Protocol Support
